Rename MaskedChatMessage file, move out of ChatMessage barrel
